controller: Move truncate method
This commit moves the `Controller::partially_truncate_status_history`
method to a function in `collection_mgmt.rs`. It is 100% code movement.

Works towards resolving MaterializeInc#28722
